# fix: slow performance when fetching user IDs repeatedly (AR-2435)