--- a/persistence/src/commonMain/kotlin/com/wire/kalium/persistence/db/UserDatabaseProvider.kt
+++ b/persistence/src/commonMain/kotlin/com/wire/kalium/persistence/db/UserDatabaseProvider.kt
@@ -11,6 +11,7 @@ import com.wire.kalium.persistence.dao.client.ClientDAO
 import com.wire.kalium.persistence.dao.message.MessageDAO
 import kotlin.jvm.JvmInline
 
+internal const val USER_CACHE_SIZE = 125
 @JvmInline
 value class UserDBSecret(val value: ByteArray)