--- a/persistence/src/commonTest/kotlin/com/wire/kalium/persistence/BaseDatabaseTest.kt
+++ b/persistence/src/commonTest/kotlin/com/wire/kalium/persistence/BaseDatabaseTest.kt
@@ -1,9 +1,11 @@
 package com.wire.kalium.persistence
 
 import com.wire.kalium.persistence.db.UserDatabaseProvider
+import kotlinx.coroutines.test.TestDispatcher
 
 expect open class BaseDatabaseTest() {
 
+    protected val dispatcher: TestDispatcher
     fun deleteDatabase()
     fun createDatabase(): UserDatabaseProvider