--- a/persistence/src/commonTest/kotlin/com/wire/kalium/persistence/cache/LRUCacheTest.kt
+++ b/persistence/src/commonTest/kotlin/com/wire/kalium/persistence/cache/LRUCacheTest.kt
@@ -1,0 +1,96 @@
+package com.wire.kalium.persistence.cache
+
+import kotlinx.coroutines.test.runTest
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class LRUCacheTest {
+
+    @Test
+    fun givenAnElementIsCached_whenGettingTheSecondTime_shouldExecuteTheProviderLambdaOnlyTheFirstTime() = runTest {
+        val cache = createCache(100)
+        val itemKey = 42
+        var providerExecCount = 0
+
+        cache.get(itemKey) {
+            providerExecCount++
+            "Number one"
+        }
+        assertEquals(1, providerExecCount)
+
+        cache.get(itemKey) { error("This should not be executed as the item is already cached") }
+        assertEquals(1, providerExecCount)
+    }
+
+    @Test
+    fun givenAKeyAndProvider_whenInvokingTheProvider_shouldPassTheKeyAsParameter() = runTest {
+        val cache = createCache(100)
+        val itemKey = 42
+        var providerExecCount = 0
+
+        cache.get(itemKey) {
+            assertEquals(itemKey, it)
+            providerExecCount++
+            "Number one"
+        }
+        assertEquals(1, providerExecCount)
+    }
+
+    @Test
+    fun givenAnElementIsCached_whenGettingItsValue_shouldReturnTheProvidedValue() = runTest {
+        val cache = createCache(100)
+        val itemKey = 42
+        val expectedElement = "I'm expected"
+
+        val result = cache.get(itemKey) { expectedElement }
+        assertEquals(expectedElement, result)
+    }
+
+    @Test
+    fun givenAnElementIsRemoved_whenGettingItsValueTheSecondTime_shouldInvokeTheProviderLambda() = runTest {
+        val cache = createCache(100)
+        val itemKey = 42
+        cache.get(itemKey) { "Zio" }
+
+        cache.remove(itemKey)
+        val secondResult = cache.get(itemKey) { "Expected Value" }
+
+        val thirdResult = cache.get(itemKey) { "Zia" }
+
+        assertEquals(secondResult, thirdResult)
+    }
+
+    @Test
+    fun givenTheCacheCountExceedsTheLimit_whenGettingAnExpiredValue_shouldInvokeTheNewProvider() = runTest {
+        val limit = 10
+        val cache = createCache(limit)
+        val itemsToExpire = 2
+        val expectedElement = "The replacement"
+
+        repeat(limit + itemsToExpire) { index ->
+            cache.get(index) { "The originals" }
+        }
+        val result = cache.get(0) { expectedElement }
+
+        assertEquals(expectedElement, result)
+    }
+
+    @Test
+    fun givenTheCacheCountExceedsTheLimit_whenGettingARecentValue_shouldNotInvokeTheNewProvider() = runTest {
+        val limit = 10
+        val cache = createCache(limit)
+        val itemsToExpire = 2
+        val expectedElement = "The originals"
+
+        repeat(limit + itemsToExpire) { index ->
+            cache.get(index) { expectedElement }
+        }
+        val result = cache.get(itemsToExpire + 1) {
+            error("Should not invoke this, as the item was not expired")
+        }
+
+        assertEquals(expectedElement, result)
+    }
+
+    private fun createCache(size: Int) = LRUCache<Int, String>(size)
+}