--- a/persistence/src/commonTest/kotlin/com/wire/kalium/persistence/dao/UserDAOTest.kt
+++ b/persistence/src/commonTest/kotlin/com/wire/kalium/persistence/dao/UserDAOTest.kt
@@ -6,6 +6,7 @@ import com.wire.kalium.persistence.db.UserDatabaseProvider
 import com.wire.kalium.persistence.utils.stubs.newUserEntity
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.take
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.TestResult
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
@@ -34,14 +35,14 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenUser_ThenUserCanBeInserted() = runTest {
+    fun givenUser_ThenUserCanBeInserted() = runTest(dispatcher) {
         db.userDAO.insertUser(user1)
         val result = db.userDAO.getUserByQualifiedID(user1.id).first()
         assertEquals(result, user1)
     }
 
     @Test
-    fun givenListOfUsers_ThenMultipleUsersCanBeInsertedAtOnce() = runTest {
+    fun givenListOfUsers_ThenMultipleUsersCanBeInsertedAtOnce() = runTest(dispatcher) {
         db.userDAO.upsertUsers(listOf(user1, user2, user3))
         val result1 = db.userDAO.getUserByQualifiedID(user1.id).first()
         val result2 = db.userDAO.getUserByQualifiedID(user2.id).first()
@@ -52,7 +53,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenExistingUser_ThenUserCanBeDeleted() = runTest {
+    fun givenExistingUser_ThenUserCanBeDeleted() = runTest(dispatcher) {
         db.userDAO.insertUser(user1)
         db.userDAO.deleteUserByQualifiedID(user1.id)
         val result = db.userDAO.getUserByQualifiedID(user1.id).first()
@@ -60,7 +61,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenExistingUser_ThenUserCanBeUpdated() = runTest {
+    fun givenExistingUser_ThenUserCanBeUpdated() = runTest(dispatcher) {
         db.userDAO.insertUser(user1)
         val updatedUser1 = UserEntity(
             user1.id,
@@ -84,35 +85,11 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenListOfUsers_ThenUserCanBeQueriedByName() = runTest {
+    fun givenRetrievedUser_ThenUpdatesArePropagatedThroughFlow() = runTest(dispatcher) {
+        val collectedValues = mutableListOf<UserEntity?>()
+
         db.userDAO.insertUser(user1)
-        val updatedUser1 = UserEntity(
-            user1.id,
-            "John Doe",
-            "johndoe",
-            "email1",
-            "phone1",
-            1,
-            "team",
-            ConnectionEntity.State.ACCEPTED,
-            UserAssetIdEntity("asset1", "domain"),
-            UserAssetIdEntity("asset2", "domain"),
-            UserAvailabilityStatusEntity.NONE,
-            UserTypeEntity.STANDARD,
-            botService = null,
-            false
-        )
 
-        val result = db.userDAO.getUserByQualifiedID(user1.id)
-        assertEquals(user1, result.first())
-
-        db.userDAO.updateSelfUser(updatedUser1)
-        assertEquals(updatedUser1, result.first())
-    }
-
-    @Test
-    fun givenRetrievedUser_ThenUpdatesArePropagatedThroughFlow() = runTest {
-        db.userDAO.insertUser(user1)
         val updatedUser1 = UserEntity(
             user1.id,
             "John Doe",
@@ -130,15 +107,18 @@ class UserDAOTest : BaseDatabaseTest() {
             false
         )
 
-        val result = db.userDAO.getUserByQualifiedID(user1.id)
-        assertEquals(user1, result.first())
-
-        db.userDAO.updateSelfUser(updatedUser1)
-        assertEquals(updatedUser1, result.first())
+        db.userDAO.getUserByQualifiedID(user1.id).take(2).collect() {
+            collectedValues.add(it)
+            if (collectedValues.size == 1) {
+                db.userDAO.updateSelfUser(updatedUser1)
+            }
+        }
+        assertEquals(user1, collectedValues[0])
+        assertEquals(updatedUser1, collectedValues[1])
     }
 
     @Test
-    fun givenAExistingUsers_WhenQueriedUserByUserEmail_ThenResultsIsEqualToThatUser() = runTest {
+    fun givenAExistingUsers_WhenQueriedUserByUserEmail_ThenResultsIsEqualToThatUser() = runTest(dispatcher) {
         // given
         val user1 = USER_ENTITY_1
         val user2 = USER_ENTITY_2.copy(email = "uniqueEmailForUser2")
@@ -162,7 +142,7 @@ class UserDAOTest : BaseDatabaseTest() {
 
     @Test
     fun givenAExistingUsers_WhenQueriedUserByName_ThenResultsIsEqualToThatUser(): TestResult {
-        return runTest {
+        return runTest(dispatcher) {
             // given
             val user1 = USER_ENTITY_1
             val user2 = USER_ENTITY_3
@@ -185,7 +165,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenAExistingUsers_WhenQueriedUserByHandle_ThenResultsIsEqualToThatUser() = runTest {
+    fun givenAExistingUsers_WhenQueriedUserByHandle_ThenResultsIsEqualToThatUser() = runTest(dispatcher) {
         // given
         val user1 = USER_ENTITY_1.copy(name = "uniqueNameFor User1")
         val user2 = USER_ENTITY_2
@@ -204,7 +184,7 @@ class UserDAOTest : BaseDatabaseTest() {
 
     @Test
     fun givenAExistingUsersWithCommonEmailPrefix_WhenQueriedWithThatEmailPrefix_ThenResultIsEqualToTheUsersWithCommonEmailPrefix() =
-        runTest {
+        runTest(dispatcher) {
             // given
             val commonEmailPrefix = "commonEmail"
 
@@ -264,7 +244,7 @@ class UserDAOTest : BaseDatabaseTest() {
 
     // when entering
     @Test
-    fun givenAExistingUsers_WhenQueriedWithNonExistingEmail_ThenReturnNoResults() = runTest {
+    fun givenAExistingUsers_WhenQueriedWithNonExistingEmail_ThenReturnNoResults() = runTest(dispatcher) {
         // given
         val mockUsers = listOf(USER_ENTITY_1, USER_ENTITY_2, USER_ENTITY_3)
         db.userDAO.upsertUsers(mockUsers)
@@ -286,7 +266,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenAExistingUsers_whenQueriedWithCommonEmailPrefix_ThenResultsUsersEmailContainsThatPrefix() = runTest {
+    fun givenAExistingUsers_whenQueriedWithCommonEmailPrefix_ThenResultsUsersEmailContainsThatPrefix() = runTest(dispatcher) {
         // given
         val commonEmailPrefix = "commonEmail"
 
@@ -310,7 +290,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenAExistingUsers_whenQueriedWithCommonHandlePrefix_ThenResultsUsersHandleContainsThatPrefix() = runTest {
+    fun givenAExistingUsers_whenQueriedWithCommonHandlePrefix_ThenResultsUsersHandleContainsThatPrefix() = runTest(dispatcher) {
         // given
         val commonHandlePrefix = "commonHandle"
 
@@ -334,7 +314,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenAExistingUsers_whenQueriedWithCommonNamePrefix_ThenResultsUsersNameContainsThatPrefix() = runTest {
+    fun givenAExistingUsers_whenQueriedWithCommonNamePrefix_ThenResultsUsersNameContainsThatPrefix() = runTest(dispatcher) {
         // given
         val commonNamePrefix = "commonName"
 
@@ -359,7 +339,7 @@ class UserDAOTest : BaseDatabaseTest() {
 
     @Test
     fun givenAExistingUsers_whenQueriedWithCommonPrefixForNameHandleAndEmail_ThenResultsUsersNameHandleAndEmailContainsThatPrefix() =
-        runTest {
+        runTest(dispatcher) {
             // given
             val commonPrefix = "common"
 
@@ -382,7 +362,7 @@ class UserDAOTest : BaseDatabaseTest() {
         }
 
     @Test
-    fun givenAExistingUsers_whenQueried_ThenResultsUsersAreConnected() = runTest {
+    fun givenAExistingUsers_whenQueried_ThenResultsUsersAreConnected() = runTest(dispatcher) {
         // given
         val commonPrefix = "common"
 
@@ -411,7 +391,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenAConnectedExistingUsersAndNonConnected_whenQueried_ThenResultsUsersAreConnected() = runTest {
+    fun givenAConnectedExistingUsersAndNonConnected_whenQueried_ThenResultsUsersAreConnected() = runTest(dispatcher) {
         // given
         val commonPrefix = "common"
 
@@ -439,7 +419,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenAConnectedExistingUserAndNonConnected_whenQueried_ThenResultIsTheConnectedUser() = runTest {
+    fun givenAConnectedExistingUserAndNonConnected_whenQueried_ThenResultIsTheConnectedUser() = runTest(dispatcher) {
         // given
         val commonPrefix = "common"
 
@@ -469,7 +449,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenTheListOfUser_whenQueriedByHandle_ThenResultContainsOnlyTheUserHavingTheHandleAndAreConnected() = runTest {
+    fun givenTheListOfUser_whenQueriedByHandle_ThenResultContainsOnlyTheUserHavingTheHandleAndAreConnected() = runTest(dispatcher) {
         val expectedResult = listOf(
             USER_ENTITY_1.copy(handle = "@someHandle"),
             USER_ENTITY_4.copy(handle = "@someHandle1")
@@ -500,7 +480,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenAExistingUsers_whenUpdatingTheirValues_ThenResultsIsEqualToThatUserButWithFieldsModified() = runTest {
+    fun givenAExistingUsers_whenUpdatingTheirValues_ThenResultsIsEqualToThatUserButWithFieldsModified() = runTest(dispatcher) {
         // given
         val newNameA = "new user naming a"
         val newNameB = "new user naming b"
@@ -517,7 +497,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenAExistingUsers_whenUpdatingTheirValuesAndRecordNotExists_ThenResultsOneUpdatedAnotherInserted() = runTest {
+    fun givenAExistingUsers_whenUpdatingTheirValuesAndRecordNotExists_ThenResultsOneUpdatedAnotherInserted() = runTest(dispatcher) {
         // given
         val newNameA = "new user naming a"
         db.userDAO.insertUser(user1)
@@ -532,7 +512,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenAExistingUsers_whenUpsertingTeamMembers_ThenResultsOneUpdatedAnotherInserted() = runTest {
+    fun givenAExistingUsers_whenUpsertingTeamMembers_ThenResultsOneUpdatedAnotherInserted() = runTest(dispatcher) {
         // given
         val newTeamId = "new user team id"
         db.userDAO.insertUser(user1)
@@ -547,7 +527,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenATeamMember_whenUpsertingTeamMember_ThenUserTypeShouldStayTheSame() = runTest {
+    fun givenATeamMember_whenUpsertingTeamMember_ThenUserTypeShouldStayTheSame() = runTest(dispatcher) {
         // given
         val externalMember = user1.copy(userType = UserTypeEntity.EXTERNAL)
         db.userDAO.upsertTeamMembersTypes(listOf(externalMember))
@@ -559,7 +539,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenAExistingUsers_whenUpsertingUsers_ThenResultsOneUpdatedAnotherInsertedWithNoConnectionStatusOverride() = runTest {
+    fun givenAExistingUsers_whenUpsertingUsers_ThenResultsOneUpdatedAnotherInsertedWithNoConnectionStatusOverride() = runTest(dispatcher) {
         // given
         val newTeamId = "new team id"
         db.userDAO.insertUser(user1.copy(connectionStatus = ConnectionEntity.State.ACCEPTED))
@@ -575,7 +555,7 @@ class UserDAOTest : BaseDatabaseTest() {
     }
 
     @Test
-    fun givenListOfUsers_WhenGettingListOfUsers_ThenMatchingUsersAreReturned() = runTest {
+    fun givenListOfUsers_WhenGettingListOfUsers_ThenMatchingUsersAreReturned() = runTest(dispatcher) {
         val users = listOf(user1, user2)
         val requestedIds = (users + user3).map { it.id }
         db.userDAO.upsertUsers(users)