--- a/persistence/src/iosX64Main/kotlin/com/wire/kalium/persistence/db/UserDatabaseProvider.kt
+++ b/persistence/src/iosX64Main/kotlin/com/wire/kalium/persistence/db/UserDatabaseProvider.kt
@@ -148,7 +148,7 @@ actual class UserDatabaseProvider(
 
     private val databaseScope = CoroutineScope(SupervisorJob() + dispatcher)
     actual val userDAO: UserDAO
-        get() = UserDAOImpl(database.usersQueries, LRUCache(100), databaseScope)
+        get() = UserDAOImpl(database.usersQueries, LRUCache(USER_CACHE_SIZE), databaseScope)
 
     actual val conversationDAO: ConversationDAO
         get() = ConversationDAOImpl(database.conversationsQueries, database.usersQueries, database.membersQueries)