--- a/persistence/src/jsTest/kotlin/com/wire/kalium/persistence/BaseDatabaseTest.kt
+++ b/persistence/src/jsTest/kotlin/com/wire/kalium/persistence/BaseDatabaseTest.kt
@@ -1,8 +1,12 @@
 package com.wire.kalium.persistence
 
 import com.wire.kalium.persistence.db.UserDatabaseProvider
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.test.StandardTestDispatcher
 
 actual open class BaseDatabaseTest actual constructor() {
+
+    protected actual val dispatcher: CoroutineDispatcher = StandardTestDispatcher()
 
     actual fun deleteDatabase() {
         // TODO delete test database