----
#### PR Submission Checklist for internal contributors

- The **PR Title**
  - [X] conforms to the style of semantic commits messages¹ supported in Wire's Github Workflow²
  - [X] contains a reference JIRA issue number like `SQPIT-764`
  - [X] answers the question: _If merged, this PR will: ..._ ³

- The **PR Description**
  - [X] is free of optional paragraphs and you have filled the relevant parts to the best of your ability
----

# What's new in this PR?

## Issues

Reloaded is querying a lot of User Details in a sequence, which contributes with DB connections deadlocking when multiple operations are being performed.

## Solutions

A simple LRU Cache to store the flows of `UserEntity` in the DAO.

This way:
- We have a single instance of `Flow<UserEntity>` for each UserId, as long as the cache isn't full.
- When the query is invalidated, we make sure that **only one** request to the DB will be made to fetch the updated details even if we have multiple observers for the same User.

As visible in `UserDAOImpl`, this `LRUCache` can be easily used for other entities if we want, like `Conversation` and `Team` for example.

## Benchmark

I've done a benchmark comparing before and after adding this in-memory LRU cache.
I've ran the same benchmark multiple times getting similar results. The result below was picked in one of the runs without any reason in particular.

<details>
  <summary>Description of the benchmark</summary>

Test Setup:
- Create 1000 users
- From these users, create a list of 10 random qualified IDs, called `idsToFetch`
- Insert all the users into the DB

In each test:
- Start measuring time
- Take one ID from the `idsToFetch` list
- Select 3 times the same user ID
- Take the next ID from the list, until all IDs were fetched 3 times.
- Stop measuring time

The code of the test, without including the setup, was this:

```kotlin
fun selectWithCaching() = runTest {
    measureTime {
        idsToFetch.forEach { userId ->
            repeat(TIMES_PER_USER) {
                measureTime {
                    userDAO.getUserByQualifiedID(userId).first()
                }.also {
                    println("Took $it to get $userId")
                }
            }
        }
    }.also {
        println(
            "Took $it to select $USERS_TO_FETCH random users from a table of $USER_COUNT; " +
                    "Selecting each user $TIMES_PER_USER times"
        )
    }
}
```

</details>


<details>
  <summary>Results WITHOUT cache</summary>

```
Took 12.694ms to get QualifiedIDEntity(value=userId18, domain=userDomain)
Took 1.638300ms to get QualifiedIDEntity(value=userId18, domain=userDomain)
Took 1.513400ms to get QualifiedIDEntity(value=userId18, domain=userDomain)
Took 1.609600ms to get QualifiedIDEntity(value=userId105, domain=userDomain)
Took 1.875300ms to get QualifiedIDEntity(value=userId105, domain=userDomain)
Took 1.546400ms to get QualifiedIDEntity(value=userId105, domain=userDomain)
Took 1.801600ms to get QualifiedIDEntity(value=userId427, domain=userDomain)
Took 1.557900ms to get QualifiedIDEntity(value=userId427, domain=userDomain)
Took 1.592900ms to get QualifiedIDEntity(value=userId427, domain=userDomain)
Took 1.504800ms to get QualifiedIDEntity(value=userId177, domain=userDomain)
Took 1.620700ms to get QualifiedIDEntity(value=userId177, domain=userDomain)
Took 1.4ms to get QualifiedIDEntity(value=userId177, domain=userDomain)
Took 1.882600ms to get QualifiedIDEntity(value=userId150, domain=userDomain)
Took 1.499800ms to get QualifiedIDEntity(value=userId150, domain=userDomain)
Took 1.518900ms to get QualifiedIDEntity(value=userId150, domain=userDomain)
Took 1.557900ms to get QualifiedIDEntity(value=userId840, domain=userDomain)
Took 1.651500ms to get QualifiedIDEntity(value=userId840, domain=userDomain)
Took 2.344800ms to get QualifiedIDEntity(value=userId840, domain=userDomain)
Took 1.829600ms to get QualifiedIDEntity(value=userId938, domain=userDomain)
Took 1.644800ms to get QualifiedIDEntity(value=userId938, domain=userDomain)
Took 1.616400ms to get QualifiedIDEntity(value=userId938, domain=userDomain)
Took 1.665ms to get QualifiedIDEntity(value=userId777, domain=userDomain)
Took 1.862ms to get QualifiedIDEntity(value=userId777, domain=userDomain)
Took 1.997100ms to get QualifiedIDEntity(value=userId777, domain=userDomain)
Took 1.813900ms to get QualifiedIDEntity(value=userId419, domain=userDomain)
Took 1.385800ms to get QualifiedIDEntity(value=userId419, domain=userDomain)
Took 1.608100ms to get QualifiedIDEntity(value=userId419, domain=userDomain)
Took 1.597300ms to get QualifiedIDEntity(value=userId209, domain=userDomain)
Took 1.927400ms to get QualifiedIDEntity(value=userId209, domain=userDomain)
Took 1.750200ms to get QualifiedIDEntity(value=userId209, domain=userDomain)
Took 66.082400ms to select 10 random users from a table of 1000; Selecting each user 3 times
```
</details>


<details>
  <summary>Results WITH cache</summary>

```
Took 68.205700ms to get QualifiedIDEntity(value=userId967, domain=userDomain)
Took 147.7us to get QualifiedIDEntity(value=userId967, domain=userDomain)
Took 147.2us to get QualifiedIDEntity(value=userId967, domain=userDomain)
Took 3.056200ms to get QualifiedIDEntity(value=userId750, domain=userDomain)
Took 198.2us to get QualifiedIDEntity(value=userId750, domain=userDomain)
Took 75.7us to get QualifiedIDEntity(value=userId750, domain=userDomain)
Took 3.936300ms to get QualifiedIDEntity(value=userId370, domain=userDomain)
Took 119us to get QualifiedIDEntity(value=userId370, domain=userDomain)
Took 78.3us to get QualifiedIDEntity(value=userId370, domain=userDomain)
Took 3.278100ms to get QualifiedIDEntity(value=userId728, domain=userDomain)
Took 70.8us to get QualifiedIDEntity(value=userId728, domain=userDomain)
Took 207.2us to get QualifiedIDEntity(value=userId728, domain=userDomain)
Took 2.750600ms to get QualifiedIDEntity(value=userId617, domain=userDomain)
Took 174.6us to get QualifiedIDEntity(value=userId617, domain=userDomain)
Took 256.6us to get QualifiedIDEntity(value=userId617, domain=userDomain)
Took 3.035900ms to get QualifiedIDEntity(value=userId366, domain=userDomain)
Took 48.9us to get QualifiedIDEntity(value=userId366, domain=userDomain)
Took 37.3us to get QualifiedIDEntity(value=userId366, domain=userDomain)
Took 3.111900ms to get QualifiedIDEntity(value=userId570, domain=userDomain)
Took 55.1us to get QualifiedIDEntity(value=userId570, domain=userDomain)
Took 196.7us to get QualifiedIDEntity(value=userId570, domain=userDomain)
Took 2.511800ms to get QualifiedIDEntity(value=userId659, domain=userDomain)
Took 51.4us to get QualifiedIDEntity(value=userId659, domain=userDomain)
Took 41.9us to get QualifiedIDEntity(value=userId659, domain=userDomain)
Took 2.454900ms to get QualifiedIDEntity(value=userId803, domain=userDomain)
Took 50.6us to get QualifiedIDEntity(value=userId803, domain=userDomain)
Took 66.7us to get QualifiedIDEntity(value=userId803, domain=userDomain)
Took 2.633500ms to get QualifiedIDEntity(value=userId565, domain=userDomain)
Took 159.8us to get QualifiedIDEntity(value=userId565, domain=userDomain)
Took 97.7us to get QualifiedIDEntity(value=userId565, domain=userDomain)
Took 110.000400ms to select 10 random users from a table of 1000; Selecting each user 3 times
```
</details>

### Conclusion

Even though there's clearly an overhead when fetching for the very first time, and a smaller but also visible overhead when fetching a new user for the first time, without cache we are not able to query faster than a low 1ms for each user.

However, with cache, subsequent calls to the same user are an order of magnitude faster, taking just a few dozens us, and it has the added benefit of **not spending DB connections**.

The applications we are developing Kalium for definitely will request the same user more than once, and it's an added benefit of using Kalium, that we can offer a multiplatform and centralised caching logic out of the box.

## Tests

### Test Coverage

- [X] I have added automated test to this contribution

## Attachments

Android Reloaded scrolls smoothly and is able to receive messages faster now.

https://user-images.githubusercontent.com/9389043/190835939-6858019e-89d5-4f9f-b21b-d53f95f6eb61.mp4

----
#### PR Post Merge Checklist for internal contributors

 - [X] If any soft of configuration variable was introduced by this PR, it has been added to the relevant documents and the CI jobs have been updated.
----
##### References
1. https://sparkbox.com/foundry/semantic_commit_messages
1. https://github.com/wireapp/.github#usage
1. E.g. `feat(conversation-list): Sort conversations by most emojis in the title #SQPIT-764`.